// src/types/Titulo.ts
export default interface Titulo {
  numNfv: number
  numTit: string
  vctPro: string
  vlrOri: number
  codEmp: number
  vlrAbe: number
  codFil: number
  numNff: number
  numPed: number
  codCli: number
  numDfs: number
  nomCli: string
  codTpt: string
  vctOri: string
  nomFil: string
  isLoading?: boolean
}